+++ crates/pdk-api/src/api/checksum.rs
@@ -6,33 +6,58 @@ use thiserror::Error;
 /// Errors that may occur from within a plugin.
 #[derive(Error, Debug)]
 pub enum ChecksumError {
-    #[error("Checksum type is not defined.")]
-    MissingChecksumType,
+    #[error("Checksum algorithm is not defined.")]
+    MissingAlgorithm,
 
-    #[error("Unknown or unsupported checksum type {kind}.")]
-    UnknownChecksumType { kind: String },
+    #[error("Unknown or unsupported checksum algorithm {kind}.")]
+    UnsupportedAlgorithm { kind: String },
+}
+
+/// Supported checksum algorithms.
+#[derive(Clone, Copy, Debug, Deserialize, PartialEq, Serialize)]
+#[serde(rename_all = "lowercase")]
+pub enum ChecksumAlgorithm {
+    Minisign,
+    Sha256,
+    Sha512,
 }
 
 /// Represents a checksum for a specific algorithm.
 #[derive(Clone, Debug, Deserialize, PartialEq, Serialize)]
 #[serde(into = "String", try_from = "String")]
-pub enum Checksum {
+pub struct Checksum {
+    /// Algorithm.
+    pub algo: ChecksumAlgorithm,
+
     /// Public key.
-    Minisign(String),
+    pub key: Option<String>,
 
     /// File hash.
-    Sha256(String),
-
-    /// File hash.
-    Sha512(String),
+    pub hash: Option<String>,
 }
 
 impl Checksum {
-    pub fn as_str(&self) -> &str {
-        match self {
-            Self::Minisign(hash) => hash,
-            Self::Sha256(hash) => hash,
-            Self::Sha512(hash) => hash,
+    pub fn minisign(key: String) -> Self {
+        Self {
+            algo: ChecksumAlgorithm::Minisign,
+            key: Some(key),
+            hash: None,
+        }
+    }
+
+    pub fn sha256(hash: String) -> Self {
+        Self {
+            algo: ChecksumAlgorithm::Sha256,
+            key: None,
+            hash: Some(hash),
+        }
+    }
+
+    pub fn sha512(hash: String) -> Self {
+        Self {
+            algo: ChecksumAlgorithm::Sha512,
+            key: None,
+            hash: Some(hash),
         }
     }
 }
@@ -43,14 +68,14 @@ impl FromStr for Checksum {
     fn from_str(value: &str) -> Result<Self, Self::Err> {
         match value.split_once(':') {
             Some((tag, hash)) => match tag {
-                "minisign" => Ok(Self::Minisign(hash.to_owned())),
-                "sha256" => Ok(Self::Sha256(hash.to_owned())),
-                "sha512" => Ok(Self::Sha256(hash.to_owned())),
-                _ => Err(ChecksumError::UnknownChecksumType {
+                "minisign" => Ok(Self::minisign(hash.to_owned())),
+                "sha256" => Ok(Self::sha256(hash.to_owned())),
+                "sha512" => Ok(Self::sha256(hash.to_owned())),
+                _ => Err(ChecksumError::UnsupportedAlgorithm {
                     kind: tag.to_owned(),
                 }),
             },
-            None => Err(ChecksumError::MissingChecksumType),
+            None => Err(ChecksumError::MissingAlgorithm),
         }
     }
 }
@@ -65,10 +90,16 @@ impl TryFrom<String> for Checksum {
 
 impl fmt::Display for Checksum {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            Self::Minisign(hash) => write!(f, "minisign:{hash}"),
-            Self::Sha256(hash) => write!(f, "sha256:{hash}"),
-            Self::Sha512(hash) => write!(f, "sha512:{hash}"),
+        match self.algo {
+            ChecksumAlgorithm::Minisign => {
+                write!(f, "minisign:{}", self.key.as_deref().unwrap_or_default())
+            }
+            ChecksumAlgorithm::Sha256 => {
+                write!(f, "sha256:{}", self.hash.as_deref().unwrap_or_default())
+            }
+            ChecksumAlgorithm::Sha512 => {
+                write!(f, "sha512:{}", self.hash.as_deref().unwrap_or_default())
+            }
         }
     }
 }